archived/jax330/requirements.txt
imageio==2.22.0
torch==1.13.0
numpy==1.23.3
torchvision==0.10.1
optax==0.0.6
jax==0.3.25
tqdm==4.64.1
easyjax==0.1.3
Pillow==9.3.0
rsbox==0.0.11
